feat: add convenience aggregation script for multiple runs
this script aggregates performance metrics and clustering scores from
several runs.

it expects the backend and run timestamp to be encoded in the output
folder, in the form:

out_BACKEND_TIMESTAMP

as produced by the Makefile in this repo.

the script can produce both csv and parquet dataframes.

it produces two different files to keep normalization of the data
structure. some of the metadata are properties of the method leaves, and
some others are properties of the metric leaves.

indexing by run ID (the timestamp), together with dataset x method x
seed x metric should be enough to JOIN the two data frames, if needed.